requirements.txt
Django==1.11.6
django-cors-headers==2.1.0
django-nested-admin==3.0.21
djangorestframework==3.7.1
djangorestframework-jwt==1.11.0
PyJWT==1.5.3
python-monkey-business==1.0.0
pytz==2017.3
six==1.11.0
